% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{coef.TFRE}
\alias{coef.TFRE}
\title{Extract coefficients from a 'TFRE' object}
\usage{
\method{coef}{TFRE}(object, s, ...)
}
\arguments{
\item{object}{Fitted "TFRE" model object.}

\item{s}{Regression model to use for coefficient extraction. Should be one of
"1st" and "2nd". See more details in "Details".}

\item{...}{Not used. Other arguments to extract coefficients.}
}
\value{
The coefficient vector from the fitted TFRE model, with the first
element as the intercept.
}
\description{
This function extracts the coefficient vector from a fitted TFRE
Lasso, SCAD or MCP model.
}
\details{
If \code{object$second_stage = "none"}, \code{s} cannot be "2nd". If
\code{object$second_stage = "none"} and \code{s = "2nd"}, the function will
return the coefficient vector from the TFRE Lasso regression. If \code{object$second_stage = "scad"}
or \code{"mcp"}, and \code{s = "2nd"}, the function will return the coefficient
vector from the TFRE SCAD or MCP regression with the smallest HBIC.
}
\examples{
n <- 100; p <- 400
beta0 <- c(1.5,-1.25,1,-0.75,0.5,rep(0,p-5))
eta_list = 0.1*8:20*sqrt(log(p)/n)
X <- matrix(rnorm(n*p),n)
y <- X \%*\% beta0 + rt(n,4)

Obj_TFRE_Lasso <- TFRE(X, y, second_stage = "none")
Obj_TFRE_SCAD <- TFRE(X, y, eta_list = eta_list)
Obj_TFRE_MCP <- TFRE(X, y, second_stage = "mcp", eta_list = eta_list)

coef(Obj_TFRE_Lasso, "1st")[1:10]
coef(Obj_TFRE_Lasso, "2nd")[1:10]
coef(Obj_TFRE_SCAD, "1st")[1:10]
coef(Obj_TFRE_SCAD, "2nd")[1:10]
coef(Obj_TFRE_MCP, "1st")[1:10]
coef(Obj_TFRE_MCP, "2nd")[1:10]

}
\references{
Wang, L., Peng, B., Bradic, J., Li, R. and Wu, Y. (2020),
\emph{A Tuning-free Robust and Efficient Approach to High-dimensional Regression,
Journal of the American Statistical Association, 115:532, 1700-1714},
\doi{10.1080/01621459.2020.1840989}.
}
\seealso{
\code{\link{TFRE}}, \code{\link{predict.TFRE}}, \code{\link{plot.TFRE}}
}
\author{
Yunan Wu and Lan Wang\cr Maintainer:
Yunan Wu <yunan.wu@utdallas.edu>
}
